Refactor eleventy mocking
- Separate context for transform and shortcode
- Updated values to reflect a typical install
