+++ testUtils/eleventyMock/index.js
@@ -1,0 +1,3 @@
+exports.mockValues = require("./values");
+exports.mockShortcode = require("./shortcode");
+exports.mockTransform = require("./transform");
