+++ testUtils/eleventyMock/values.js
@@ -1,0 +1,9 @@
+// Common mocking values
+// Based on typical default 11ty setup
+
+exports.inputPath = "./src/index.md";
+exports.inputDir = "src";
+exports.outputDir = "_site";
+exports.outputPath = "_site/index.html";
+exports.filePathStem = "/index";
+exports.url = "/";
